move all send_notification tests to separate file
